Деплой телеграм-бота в интернет с помощью сервиса heroku. Смена IDE (Eclipse => IntelliJ IDEA).
