--- .classpath
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER">
-		<attributes>
-			<attribute name="module" value="true"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="con" path="org.eclipse.jdt.junit.JUNIT_CONTAINER/5"/>
-	<classpathentry kind="lib" path="lib/telegrambots-4.1-jar-with-dependencies.jar"/>
-	<classpathentry kind="lib" path="lib/telegrambots-chat-session-bot-4.1-jar-with-dependencies.jar"/>
-	<classpathentry kind="lib" path="lib/TelegramBots-4.1.zip"/>
-	<classpathentry kind="lib" path="lib/telegrambots-abilities-4.1-jar-with-dependencies.jar"/>
-	<classpathentry kind="lib" path="lib/telegrambots-meta-4.1-jar-with-dependencies.jar"/>
-	<classpathentry kind="lib" path="lib/telegrambots-spring-boot-starter-4.1-jar-with-dependencies.jar"/>
-	<classpathentry kind="lib" path="lib/telegrambotsextensions-4.1-jar-with-dependencies.jar"/>
-	<classpathentry kind="lib" path="lib/telegrambots-4.1.jar"/>
-	<classpathentry kind="lib" path="lib/telegrambots-abilities-4.1.jar"/>
-	<classpathentry kind="lib" path="lib/telegrambots-chat-session-bot-4.1.jar"/>
-	<classpathentry kind="lib" path="lib/telegrambots-meta-4.1.jar"/>
-	<classpathentry kind="lib" path="lib/telegrambots-spring-boot-starter-4.1.jar"/>
-	<classpathentry kind="lib" path="lib/telegrambotsextensions-4.1.jar"/>
-	<classpathentry kind="lib" path="lib/sdk-0.5.12.jar"/>
-	<classpathentry kind="lib" path="lib/gson-2.2.2.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
