--- .settings/org.eclipse.core.resources.prefs
@@ -1,5 +1,0 @@
-eclipse.preferences.version=1
-encoding//src/First/Bot.java=UTF-8
-encoding//src/First/Hangman.java=UTF-8
-encoding/<project>=UTF-8
-encoding/README.md=UTF-8
