--- src/First/ConsoleVersion.java
@@ -1,35 +1,0 @@
-package First;
-
-import java.util.Scanner;
-
-import java.util.HashMap;
-
-public class ConsoleVersion {
-	private static Scanner scanner = new Scanner(System.in);
-	private static User currentUser = new User(0);
-	
-	public static void main(String[] args) {
-		var bot = GeneratorBot.getBot();
-		System.out.println(Bot.start());
-		currentUser.addListener(bot);
-		processInput();
-	}
-
-	private static void processInput() {
-		while (true) {
-			if(scanner.hasNextLine()) {			
-				var messageFromUser = scanner.nextLine();
-				currentUser.sendMessage(messageFromUser);
-				printBotMessages();
-			}
-		}
-	}
-
-	private static void printBotMessages() {
-		for (var messageFromBot : currentUser.getReceivedFromBotMessages()) {
-			if (messageFromBot != null) {
-				System.out.println(messageFromBot);
-			}
-		}
-	}
-}
