--- src/First/Game.java
@@ -1,8 +1,0 @@
-package First;
-
-public interface Game {
-	public String start();
-	public boolean getGameIsOver();
-	public String getName();
-	public String play(User user);
-}
