--- src/First/MessageListener.java
@@ -1,5 +1,0 @@
-package First;
-
-public interface MessageListener {
-	String onMessage(String message, User currentUser);
-}
