+++ src/main/java/First/Game.java
@@ -1,0 +1,8 @@
+package First;
+
+public interface Game {
+	public String start();
+	public boolean getGameIsOver();
+	public String getName();
+	public String play(User user);
+}
