+++ src/main/java/First/MessageListener.java
@@ -1,0 +1,5 @@
+package First;
+
+public interface MessageListener {
+	String onMessage(String message, User currentUser);
+}
