+++ system.properties
@@ -1,0 +1,1 @@
+java.runtime.version=10
